move sqlite inside database directory
