+++ be1-go/internal/database/sqlite/sqlite_const.go
