+++ be1-go/internal/hub/hub.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"go.dedis.ch/kyber/v3"
 	"popstellar/internal/crypto"
+	"popstellar/internal/database/sqlite"
 	"popstellar/internal/errors"
 	answerHandler "popstellar/internal/handler/answer"
 	"popstellar/internal/handler/jsonrpc"
@@ -30,7 +31,6 @@ import (
 	"popstellar/internal/message/query/method"
 	"popstellar/internal/network/socket"
 	"popstellar/internal/repository"
-	"popstellar/internal/sqlite"
 	"popstellar/internal/types"
 	"popstellar/internal/validation"
 	"time"
